# control-plane/rest/openapi-specs/v0_api_spec.yaml
openapi: 3.0.3
info:
  title: Mayastor RESTful API
  version: v0
servers:
  - url: /v0
paths:
  /nexuses:
    get:
      tags:
        - Nexuses
      operationId: get_nexuses
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Nexus"
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
  "/nexuses/{nexus_id}":
    get:
      tags:
        - Nexuses
      operationId: get_nexus
      parameters:
        - in: path
          name: nexus_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Nexus"
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
    delete:
      tags:
        - Nexuses
      operationId: del_nexus
      parameters:
        - in: path
          name: nexus_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: OK
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
...
  "/nodes/{id}/nexuses":
    get:
      tags:
        - Nexuses
      operationId: get_node_nexuses
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Nexus"
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
...
  "/nodes/{node_id}/nexuses/{nexus_id}":
    get:
      tags:
        - Nexuses
      operationId: get_node_nexus
      parameters:
        - in: path
          name: node_id
          required: true
          schema:
            type: string
        - in: path
          name: nexus_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Nexus"
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
    put:
      tags:
        - Nexuses
      operationId: put_node_nexus
      parameters:
        - in: path
          name: node_id
          required: true
          schema:
            type: string
        - in: path
          name: nexus_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNexusBody"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Nexus"
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
    delete:
      tags:
        - Nexuses
      operationId: del_node_nexus
      parameters:
        - in: path
          name: node_id
          required: true
          schema:
            type: string
        - in: path
          name: nexus_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: OK
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
...
  "/nodes/{node_id}/nexuses/{nexus_id}/share":
    delete:
      tags:
        - Nexuses
      operationId: del_node_nexus_share
      parameters:
        - in: path
          name: node_id
          required: true
          schema:
            type: string
        - in: path
          name: nexus_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: OK
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
  "/nodes/{node_id}/nexuses/{nexus_id}/share/{protocol}":
    put:
      tags:
        - Nexuses
      operationId: put_node_nexus_share
      parameters:
        - in: path
          name: node_id
          required: true
          schema:
            type: string
        - in: path
          name: nexus_id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: protocol
          required: true
          schema:
            $ref: "#/components/schemas/NexusShareProtocol"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
...
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ChildState:
      example: Online
      description: State of a Nexus Child
      type: string
      enum:
        - Unknown
        - Online
        - Degraded
        - Faulted
    Child:
      example:
        rebuildProgress: ~
        state: Online
        uri: "nvmf://10.1.0.5:8420/nqn.2019-05.io.openebs:37d83441-e8ef-4e17-a29e-25169d91cb96"
      description: Child information
      type: object
      properties:
        rebuildProgress:
          description: current rebuild progress (%)
          type: integer
          minimum: 0
        state:
          description: state of the child
          allOf:
            - $ref: "#/components/schemas/ChildState"
        uri:
          description: uri of the child device
          type: string
      required:
        - state
        - uri
    CreateNexusBody:
      example:
        children:
          - "nvmf://10.1.0.5:8420/nqn.2019-05.io.openebs:37d83441-e8ef-4e17-a29e-25169d91cb96"
        size: 80241024
      description: Create Nexus Body JSON
      type: object
      properties:
        children:
          description: "replica can be iscsi and nvmf remote targets or a local spdk bdev\n (i.e. bdev:///name-of-the-bdev).\n\n uris to the targets we connect to"
          type: array
          items:
            type: string
        size:
          description: size of the device in bytes
          type: integer
          format: int64
          minimum: 0
      required:
        - children
        - size
...
    NexusState:
      description: State of the Nexus
      type: string
      enum:
        - Unknown
        - Online
        - Degraded
        - Faulted
    Nexus:
      example:
        children:
          - rebuildProgress: ~
            state: Online
            uri: "nvmf://10.1.0.5:8420/nqn.2019-05.io.openebs:replica1"
        deviceUri: ~
        node: "ksnode-1"
        rebuilds: 0
        share: "nvmf"
        size: 8024024
        state: Online
        uuid: 514ed1c8-7174-49ac-b9cd-ad44ef670a67
      description: Nexus information
      type: object
      properties:
        children:
          description: Array of Nexus Children
          type: array
          items:
            $ref: "#/components/schemas/Child"
        deviceUri:
          description: "URI of the device for the volume (missing if not published).\n Missing property and empty string are treated the same."
          type: string
        node:
          description: id of the mayastor instance
          type: string
        rebuilds:
          description: total number of rebuild tasks
          type: integer
          minimum: 0
        share:
          $ref: "#/components/schemas/Protocol"
        size:
          description: size of the volume in bytes
          type: integer
          format: int64
          minimum: 0
        state:
          $ref: "#/components/schemas/NexusState"
        uuid:
          description: uuid of the nexus
          type: string
          format: uuid
      required:
        - children
        - deviceUri
        - node
        - rebuilds
        - share
        - size
        - state
        - uuid
...
    RestJsonError:
      example:
        details: "The Pool 'pooloop' was not found"
        kind: NotFound
      description: Rest Json Error format
      type: object
      properties:
        details:
          description: detailed error information
          type: string
        kind:
          description: error kind
          type: string
          enum:
            - Timeout
            - Deserialize
            - Internal
            - InvalidArgument
            - DeadlineExceeded
            - NotFound
            - AlreadyExists
            - PermissionDenied
            - ResourceExhausted
            - FailedPrecondition
            - NotShared
            - NotPublished
            - AlreadyPublished
            - AlreadyShared
            - Aborted
            - OutOfRange
            - Unimplemented
            - Unavailable
            - Unauthenticated
            - Unauthorized
            - Conflict
            - FailedPersist
            - Deleting
      required:
        - details
        - kind
...
    NexusShareProtocol:
      description: Nexus Share Protocol
      type: string
      enum:
        - nvmf
        - iscsi
...
    Protocol:
      description: Common Protocol
      type: string
      enum:
        - none
        - nvmf
        - iscsi
        - nbd
...